+++ lib/widgets/journal/editor/editor_widget.dart
@@ -57,6 +57,10 @@ class EditorWidget extends StatelessWidget {
           }
         }
 
+        if (!snapshot.showEditor) {
+          return const SizedBox.shrink();
+        }
+
         return RawKeyboardListener(
           focusNode: FocusNode(),
           onKey: (RawKeyEvent event) {