+++ test/widgets/journal/entry_details/delete_icon_widget_test.dart
@@ -27,6 +27,7 @@ void main() {
           entryId: testTextEntry.meta.id,
           entry: testTextEntry,
           showMap: false,
+          showEditor: true,
         ),
       );