+++ test/widgets/journal/entry_details/entry_detail_footer_test.dart
@@ -40,6 +40,7 @@ void main() {
           entryId: testTextEntry.meta.id,
           entry: testTextEntry,
           showMap: false,
+          showEditor: true,
         ),
       );
     });
@@ -136,6 +137,7 @@ void main() {
           entryId: testEntry.meta.id,
           entry: testEntry,
           showMap: false,
+          showEditor: true,
         ),
       );
       Future<void> mockStartTimer() => mockTimeService.start(testEntry);
@@ -184,6 +186,7 @@ void main() {
           entryId: testEntry.meta.id,
           entry: testEntry,
           showMap: false,
+          showEditor: true,
         ),
       );