+++ test/widgets/journal/entry_details/entry_detail_header_test.dart
@@ -33,6 +33,7 @@ void main() {
           entryId: testTextEntry.meta.id,
           entry: testTextEntry,
           showMap: false,
+          showEditor: true,
         ),
       );
     });
@@ -109,6 +110,7 @@ void main() {
           entryId: testTextEntry.meta.id,
           entry: testTextEntry,
           showMap: false,
+          showEditor: true,
         ),
       );
 
@@ -133,6 +135,7 @@ void main() {
           entryId: testTextEntry.meta.id,
           entry: testTextEntry,
           showMap: false,
+          showEditor: true,
         ),
       );
 
@@ -164,6 +167,7 @@ void main() {
           entryId: testTextEntry.meta.id,
           entry: testTextEntry.copyWith(geolocation: null),
           showMap: false,
+          showEditor: true,
         ),
       );
 
@@ -190,6 +194,7 @@ void main() {
           entryId: testTextEntry.meta.id,
           entry: testTextEntry,
           showMap: false,
+          showEditor: true,
         ),
       );