+++ test/widgets/journal/tags/tag_add_test.dart
@@ -49,6 +49,7 @@ void main() {
           entryId: testTextEntry.meta.id,
           entry: testTextEntry,
           showMap: false,
+          showEditor: true,
         ),
       );